implemented ++ and --

+++ main.go
@@ -46,7 +46,7 @@ func main() {
 }
 
 func Main(args []string) {
-	m := make(map[int]int)
+	m := make(map[float32]float32)
 	(m[0]) = 1
 
 	env := gmi.NewEnv(nil, "builtin")
